pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "stnbeta"
version = "0.1.0"
description = "STN beta-burst detection and DYNAP-SE1 encoding"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "mne>=1.0",
    "joblib",
    "psutil",
]

[project.optional-dependencies]
bursts = ["specparam", "fooof", "seaborn", "pyarrow"]
phase4 = ["matplotlib", "seaborn", "pyarrow", "PyYAML"]
phase4-cpu = ["brian2"]
phase4-gpu = ["brian2cuda"]
snn    = ["torch", "rockpool"]
dev    = ["pytest", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
